refactor: update chat bounded context to use VoteStatus from domain package
