+++ src/main/java/com/ject/vs/chat/adapter/web/ChatController.java
@@ -3,7 +3,7 @@ package com.ject.vs.chat.adapter.web;
 import com.ject.vs.chat.adapter.web.dto.*;
 import com.ject.vs.chat.port.in.*;
 import com.ject.vs.chat.port.in.dto.*;
-import com.ject.vs.vote.port.in.dto.VoteStatus;
+import com.ject.vs.vote.domain.VoteStatus;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
